mvc works, backend works

+++ BTCChart/ASPNETPT/ASPNETPT/Models/BTCprop.cs
@@ -10,14 +10,13 @@ namespace ASPNETPT.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Rate { get; set; }
+        public double Rate { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public decimal Ask { get; set; }
-        public decimal Bid { get; set; }
+        public double Ask { get; set; }
+        public double Bid { get; set; }
 
-        //public IQueryable<BtCprop> query { get; set; }
-        //public ICollection<BtCprop> Btcs { get; set; }
+        //public ICollection<BtCprop> Btc { get; set; }
 
     }
 }
