+++ BTCChart/BCCCBackend/BCCCBackend/BtcProp.cs
@@ -12,11 +12,11 @@ namespace BCCCBackend
     class BtcProp
     {
         public string Name { get; set; }
-        public decimal Rate { get; set; }
+        public double Rate { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public decimal Ask { get; set; }
-        public decimal Bid { get; set; }
+        public double Ask { get; set; }
+        public double Bid { get; set; }
         
         
     }
